I'm cleaning things up before a release.  See the release notes.  A significant change here is to the use of Makefile.master and the configure scripts (not proper GNU configure scripts, but scripts that use sed to make the machine specific makefiles).
git-svn-id: https://svn.aoc.nrao.edu/repos/sched/trunk/sched@142 d0039845-cc32-0410-9ae4-e027caa9de09

+++ src/Plot/plaxck.f
@@ -6,7 +6,7 @@ C
       INCLUDE 'plot.inc'
 C
       CHARACTER     CH, STRVAL*14
-      INTEGER*8     LMIN, LMAX
+      INTEGER       LMIN, LMAX
       INTEGER       I, AX, EL, K, OFS
       REAL          X1, X2, Y1, Y2
 C ----------------------------------------------------------------------
